Download an archive of the All Countries (all data) table

The allcountries file is a reference file of the Geonames daily data dump from 2018-01-01 containing 11,180,293 georeferenced place names. The Geonames daily data dump for all countries (Geonames allCountries.txt). The package reference file from 2018-01-01 contains 11,180,293 georeferenced place names in a 364MB file. As this is a large file it can take a while. To import the most recent file use places_import.

places_all()

Format

A data.frame with 20 variables

geonameid

name

asciiname

alternatenames

latitude

longitude

feature_class

feature_code

iso

cc2

admin1_code

admin2_code

admin3_code

admin4_code

population

elevation

dem

timezone

modification_date

download_date

Source

(Geonames allCountries file)

Details

Column names converted to lower case, punctuation and spaces replaced with underscores. The download_date column is added by places_import.

References

Citation: Geonames AllCountries (http://download.geonames.org/export/dump/allCountries.zip).

Examples

# NOT RUN {
places <- places()
# }