Import the latest Countryinfo file

The geonames country info text file http://download.geonames.org/export/dump/countryInfo.txt contains the country names and a range of codes along with information linked to the capital city.

places_countryinfo()

Value

data.frame

Details

Column names are converted to lowercase, spaces are replaced with underscores and brackets removed.

Examples

{places_countryinfo()}

#> # A tibble: 252 x 19
#>    iso   iso3  isonumeric fips  country      capital   area_in_sq_km population
#>    <chr> <chr> <chr>      <chr> <chr>        <chr>             <dbl>      <int>
#>  1 AD    AND   020        AN    Andorra      Andorra …          468.      84000
#>  2 AE    ARE   784        AE    United Arab… Abu Dhabi        82880.    4975593
#>  3 AF    AFG   004        AF    Afghanistan  Kabul           647500.   29121286
#>  4 AG    ATG   028        AC    Antigua and… St. John…          443.      86754
#>  5 AI    AIA   660        AV    Anguilla     The Vall…          102.      13254
#>  6 AL    ALB   008        AL    Albania      Tirana           28748.    2986952
#>  7 AM    ARM   051        AM    Armenia      Yerevan          29800.    2968000
#>  8 AO    AGO   024        AO    Angola       Luanda         1246700.   13068161
#>  9 AQ    ATA   010        AY    Antarctica   <NA>          14000000.          0
#> 10 AR    ARG   032        AR    Argentina    Buenos A…      2766890.   41343201
#> # ... with 242 more rows, and 11 more variables: continent <chr>, tld <chr>,
#> #   currencycode <chr>, currencyname <chr>, phone <chr>,
#> #   postal_code_format <chr>, postal_code_regex <chr>, languages <chr>,
#> #   geonameid <int>, neighbours <chr>, equivalentfipscode <chr>